Разделил страницу editor по FSD: create-thread отвечает за model/ui формы, а create-thread-editor композирует форму с text-editor без cross-slice зависимости
