+++ frontend/src/features/create-thread/index.ts
@@ -1,0 +1,1 @@
+export { CreateThreadForm } from './ui/CreateThreadForm';
