+++ frontend/src/widgets/create-thread-editor/ui/CreateThreadEditor.tsx
@@ -1,0 +1,12 @@
+'use client';
+
+import { CreateThreadForm } from '@/features/create-thread';
+import { TextEditor } from '@/features/text-editor';
+
+export function CreateThreadEditor() {
+  return (
+    <CreateThreadForm>
+      <TextEditor />
+    </CreateThreadForm>
+  );
+}
