+++ frontend/src/widgets/create-thread-editor/ui/index.ts
@@ -1,0 +1,1 @@
+export { CreateThreadEditor } from './CreateThreadEditor';
